requirements.txt
apipkg==1.5
argcomplete==1.9.4
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
check-manifest==0.37
coverage==4.5.3
devpi-client==4.3.0
devpi-common==3.3.1
entrypoints==0.3
execnet==1.5.0
filelock==3.0.10
flake8==3.7.7
idna==2.8
mccabe==0.6.1
meross-iot==0.1.4.3
more-itertools==6.0.0
paho-mqtt==1.4.0
pep8==1.7.1
pkginfo==1.5.0.1
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.3.1
pytest-cache==1.0
pytest-cov==2.6.1
pytest-mccabe==0.1
pytest-pep8==1.0.6
requests==2.21.0
six==1.12.0
toml==0.10.0
tox==3.7.0
urllib3==1.24.1
virtualenv==16.4.3
